api/requirements.txt
azure-functions
openai>=1.12.0
pypdf>=4.2.0
python-docx>=1.1.2
